refactor: move ChallengeAiModal component to challenge-ai-modal directory
