+++ src/features/challenge-ai-modal/__tests__/ChallengeAiModal.test.tsx
